refactor(gcal_setup): refactor gcal setup functions into another file
